pyproject.toml
[project]
name = "eval-rangefinder"
version = "0.1.0"
description = "Computer vision pipeline quality evaluation for the rangefinder system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "numpy>=1.24",
    "rich>=13.0",
    "questionary>=2.0",
    "gspread>=6.0",
    "google-auth>=2.0",
]

[project.scripts]
eval-rangefinder = "eval_rangefinder.tui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eval_rangefinder"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
